+++ packages/ai-constructs/src/conversation/runtime/bedrock_converse_adapter.ts
@@ -239,7 +239,7 @@ export class BedrockConverseAdapter {
               }
             } else if (chunk.contentBlockDelta.delta?.text) {
               text += chunk.contentBlockDelta.delta.text;
-              yield {
+              const amplifyChunk: StreamingResponseChunk = {
                 accumulatedTurnContent: [...accumulatedTurnContent, { text }],
                 conversationId: this.event.conversationId,
                 associatedUserMessageId: this.event.currentMessageId,
@@ -247,6 +247,14 @@ export class BedrockConverseAdapter {
                 contentBlockIndex: blockIndex,
                 contentBlockDeltaIndex: blockDeltaIndex,
               };
+              // padding is sent from Bedrock but not included in the API.
+              if ('p' in chunk.contentBlockDelta) {
+                const bedrockPadding = chunk.contentBlockDelta.p;
+                if (typeof bedrockPadding === 'string') {
+                  amplifyChunk.p = bedrockPadding;
+                }
+              }
+              yield amplifyChunk;
               lastBlockDeltaIndex = blockDeltaIndex;
               blockDeltaIndex++;
             }