+++ packages/sandbox/src/file_watching_sandbox.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, it, mock } from 'node:test';
 import path from 'path';
-import watcher from '@parcel/watcher';
+import watcher, { subscribe as _subscribe } from '@parcel/watcher';
 import {
   CDK_DEFAULT_BOOTSTRAP_VERSION_PARAMETER_NAME,
   FileWatchingSandbox,
@@ -40,7 +40,11 @@ import { EOL } from 'os';
 
 // Watcher mocks
 const unsubscribeMockFn = mock.fn();
-const subscribeMock = mock.method(watcher, 'subscribe', async () => {
+const subscribeMock = mock.fn<
+  (
+    ...args: Parameters<typeof _subscribe>
+  ) => Promise<{ unsubscribe: typeof unsubscribeMockFn }>
+>(async () => {
   return { unsubscribe: unsubscribeMockFn };
 });
 const packageManagerControllerFactory = new PackageManagerControllerFactory(
@@ -148,7 +152,8 @@ void describe('Sandbox to check if region is bootstrapped', () => {
       ssmClientMock,
       functionsLogStreamerMock as unknown as LambdaFunctionLogStreamer,
       printer as unknown as Printer,
-      openMock as never
+      openMock as never,
+      subscribeMock as never
     );
 
     ssmClientSendMock.mock.resetCalls();
@@ -1172,7 +1177,8 @@ const setupAndStartSandbox = async (
     testData.ssmClient,
     testData.functionsLogStreamer,
     printer as unknown as Printer,
-    testData.open ?? _open
+    testData.open ?? _open,
+    subscribeMock as never
   );
 
   await sandboxInstance.start(sandboxOptions);